// archive/ReferenceVsCodeExtensions/src/enums.ts
export enum ElementTypes {
    server = "Server",
    deployToServer = "DeployToServer",
    connectors = "Connectors",
    connector = "Connector",
    workflows = "Workflows",
    workflow = "Workflow",
    error = "Error",
    errors = "Errors",
    info = "Info",
    infos = "Infos",
    favorites = "Favorites",
    favoriteItem = "FavoriteItem",
    workTypes = "WorkTypes",
    workType = "WorkType",
    form = "Form",
    folder = "Folder",
    folderItem = "FolderItem",
    object = "object",
    ide = "IDE",
    css = "Css",
    html = "Html",
    js = "Js",
    json = "json",
    text = "Text",
    wfActionManifest = "WfActionManifest",
    widgetManifest = "WidgetManifest",
    workflowDefinition = "workflowDefinition",
    workflowSteps = "workflowSteps",
    workflowVariables = "workflowVariables",
    workflowStep = "workflowStep",
    workflowStepActions = "workflowStepActions",
    workflowStepIDEData = "workflowStepIDEData",
    workflowStepActionConfig = "workflowStepActionConfig",
    workflowStepActionConnections = "workflowStepActionConnections",
    workflowStepAction = "workflowStepAction",
    createPermissions = "createPermissions",
    permission = "permission",
    workTypeCreatePermissions = "workTypeCreatePermissions",
    workTypeCreatePermission = "workTypeCreatePermission",
    workTypeParticipantRole = "workTypeParticipantRole",
    workTypeParticipantRolePermission = "workTypeParticipantRolePermission",
    workTypeParticipantRoles = "workTypeParticipantRoles",
    workTypeAspect = "workTypeAspect",
    workTypeAspects = "workTypeAspects",
    workTypeAspectsSections = "workTypeAspectsSections",
    workTypeAspectSection = "workTypeAspectSection",
    workTypeAspectConfig = "workTypeAspectConfig",
    optionSet = "optionSet",
    optionSets = "optionSets",
    forms = "forms",
    formField = "formField",
    // Execution Engine Types
    executionOverview = "executionOverview",
    executingPlans = "executingPlans",
    executingPlan = "executingPlan",
    executionStat = "executionStat",
    planDetail = "planDetail",
    subProcesses = "subProcesses",
    subProcess = "subProcess",
    advisorIssues = "advisorIssues",
    advisorIssue = "advisorIssue",
    issueGroup = "issueGroup"
}